fix: CropImage EF - explicit FK mapping for crop_id and user_id

+++ src/AgroApp.Application/Features/CropImages/DTOs/CropImageDto.cs
@@ -8,7 +8,7 @@ public record CropImageDto(
     string StorageKey,
     string? Category,
     string? AiDiagnosis,
-    decimal? AiConfidence,
+    float? AiConfidence,
     DateTime? TakenAt,
     DateTime CreatedAt
 );
